pyproject.toml
[project]
name = "rpa-tennis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.2",
    "selenium>=4.43.0",
    "webdriver-manager>=4.0.2",
]
